+++ src/subtree/openbmc/pldm/libpldm/CHANGELOG.md
@@ -29,3 +29,4 @@ Change categories:
 
 1. requester: Make pldm_open() return existing fd
 2. transport: Prevent sticking in waiting for response
+3. transport: Match on response in pldm_transport_send_recv_msg()